updated notes with annoying details about postgres
Seriously, postgres is really weird. The .pgpass file is NOT a config file like .my.cnf; it literally just pattern-matches for host/post/db/user combos to decide what password to send. It CANNOT be used to set anything like username or host or db or port. It is NOT a config file.

Also, PGUSER does not seem to make any difference at all. I've tried setting it a variety of ways, and psql seems to use my unix username no matter what. I finally just created a postgres account for my unix username, and a corresponding db with the same name, just so that I can log in more easily. Because postgres also requires that you connect to a database, which defaults to the username, and also cannot be changed with the shell variable PGDATABASE.

Finally, I spent a lot of non-git-visible time installing postgres 9.6, which meant that I needed to do heroku pg:pull to copy the remote production db to local for testing. This sucked because apparently I had changed the number of a migration and this was screwing things up. I still don't quite understand why but there are notes in Notes.md. This should be fixed at some point.