fixed course parsing
Can now update a question with new response information without changing the correct answers (useful if we load a new course and there is yet another way to store vote counts/percentages for a question that requires another regexp change).

Can specify a list of clicker IDs to be removed from the aggregate counts, and we can use the same upadte mechanisms to accomplish this. Useful if IRBs require permission that not all students grant, or students are under 18, and so on.

Improved command_line.rb to use DATABASE_URL if available, and otherwise to use the local DB as specified in config/database.yml.

Split commands to call all the parsers from parse.rb into run_parser.rb.