refactor: class_periods controller for modified+ extended match category updates
Using the fact that in activerecord, you can use model[field] where field is the string name of a field in order to access the data at the field. This lets us loop over a list of fields, which means we can simplify some of the update code rather than cut-paste.