turning off ajax calls due to performance bottleneck
The AJAX calls are filling up all of the memory. They are inefficient and somehow should be improved with a join or something like that.

Not sure why after the AJAX call fails, the memory does not seem to be garbage collected on Heroku. Oh well; I guess it's one of those weird dynamic performance issues.